Refactored and added tests
