+++ tests/unit/landingai/pipeline/test_image_source.py
@@ -3,9 +3,16 @@ from pathlib import Path
 from typing import List
 
 import PIL.Image
+import cv2
+import numpy as np
 import pytest
 
-from landingai.pipeline.image_source import ImageFolder
+from landingai.pipeline.image_source import (
+    ImageFolder,
+    NetworkedCamera,
+    probe_video,
+    sample_images_from_video,
+)
 from landingai.pipeline.frameset import FrameSet
 
 
@@ -88,3 +95,66 @@ def _create_file_under(tmp_dir, sub_dir, name) -> Path:
 @pytest.fixture
 def input_image_folder(input_image_files) -> Path:
     return input_image_files[0].parent
+
+
+# TODO: solve the problem of exit code 134 when running the following test in GitHub Actions
+# @patch("landingai.io.cv2.waitKey")
+# @patch("landingai.io.cv2.VideoCapture")
+# def test_read_from_notebook_webcam(mock_video_capture, mock_wait_key):
+#     mock_video_capture.return_value.read.return_value = (True, np.zeros((480, 640, 3)))
+#     mock_wait_key.return_value = 288
+#     take_photo_func = read_from_notebook_webcam()
+#     filepath = take_photo_func()
+#     image = PIL.Image.open(filepath)
+#     assert image.size == (640, 480)
+
+
+def test_sample_images_from_video(tmp_path: Path):
+    test_video_file_path = "tests/data/videos/test.mp4"
+    result = sample_images_from_video(test_video_file_path, tmp_path)
+    assert len(result) == 2
+    assert len(list(tmp_path.glob("*.jpg"))) == 2
+
+
+def test_probe():
+    test_video_file_path = "tests/data/videos/test.mp4"
+    total_frames, sample_size, video_length_seconds = probe_video(
+        test_video_file_path, 1.0
+    )
+    assert total_frames == 48
+    assert sample_size == 2
+    assert video_length_seconds == 2.0
+
+
+def test_probe_file_not_exist(tmp_path: Path):
+    with pytest.raises(FileNotFoundError):
+        non_exist_file = str(tmp_path / "non_exist.mp4")
+        probe_video(non_exist_file, 1.0)
+
+
+def test_networked_camera():
+    # Use a video to simulate a live camera and test motion detection
+    test_video_file_path = "tests/data/videos/countdown.mp4"
+    Camera = NetworkedCamera(
+        stream_url=test_video_file_path, motion_detection_threshold=50
+    )
+
+    # Get the first frame and the next frame where motion is detected. I keep the detection threshold low to make the test fast
+    i = iter(Camera)
+    frame1 = next(i)
+    while True:
+        # if we cannot get any motion detection (e.g. Threshold 100%), next() will throw an exception and fail the test
+        frame2 = next(i)
+        if not frame2.is_empty():
+            break
+    # frame1.show_image()
+    # frame2.show_image()
+    image_distance = np.sum(
+        cv2.absdiff(src1=frame1[0].to_numpy_array(), src2=frame2[0].to_numpy_array())
+    )
+
+    # Compute the diff by summing the delta between each pixel across the two images
+    del Camera
+    assert (
+        image_distance > 100000
+    )  # Even with little motion this number should exceed 100k
